move string stuff to new file
