+++ src/main/kotlin/io/github/kjm015/kylenewer/commands/ComputerCommand.kt
@@ -37,16 +37,3 @@ class ComputerCommand : Command() {
                 "${motherboard.toString().withArticle()}.")
     }
 }
-
-fun String.withArticle(): String {
-    return if (
-            this.startsWith(prefix = "a", ignoreCase = true) ||
-            this.startsWith(prefix = "e", ignoreCase = true) ||
-            this.startsWith(prefix = "i", ignoreCase = true) ||
-            this.startsWith(prefix = "o", ignoreCase = true) ||
-            this.startsWith(prefix = "u", ignoreCase = true)
-    )
-        "an $this"
-    else
-        "a $this"
-}
